misc: generalized parameter processing
